test(mo2ldap): [#64963] test rolebinding mapping
This commit introduces a test to ensure that rolebindings can be mapped
correctly to role-specific groups in LDAP via `mo_to_ldap` mapping
templates.

To achieve the mapping it was necessary to introduce a template
`role_uuid_to_person_uuids`, which maps from a role-class to the people
who has a rolebinding with the given class.