feat(graphql): [#65186] introduce to_response_list result translator
This commit introduces the `to_response_list` result translator which
converts a resolver response to its corresponding list of `Response`
objects enabling us to slowly convert our API into one which returns
`Response` objects for all relations.

The goal of returning Response objects for all relations is to gracefully
support temporal foreign-key relations in which the related object has
multiple active validities within the interval of the pointees interval.

In the future a non-list version; `to_response` should be introduced,
but this should be somewhat simple as its implementation will be
identical to the current one, except that the non-list version will wrap
the current version in a `one` or `only` call as needed.