test(graphql): [#65186] introduce validity test
This commit introduces a test to show an issue with the current
implementation of the roles field, specifically that the UUID of the
`Response` object is only set if the role is current.

This is a bug, as the user should be able to read any validity using the
`validities` field on the `Response` object, thus a `Response` object
should always be returned irrespective of the validities at play.