test(conftest): [#65186] add create_facet and role_facet fixtures
This commit introduces the `create_facet` and `role_facet` fixtures as
they will be useful in refactoring the `test_itsystem_roles.py` tests
into depending on the `empty_db` fixture in favor of the `fixture_db`
fixture.

We generally prefer tests to explicitly declare their dependencies
instead of depending on a bulk-load of fixture data.