fix(environment): [#65510] fix temporally spread addresses
This commit fixes an issue where temporally spread addresses would be
treated as if they were overlapping in the `load_org_unit_address`
function, making the function unable to handle multiple subsequent
addresses effectively breaking on production data as produced by the pay
system.

The issue was introduced in version 35.1.1 of the integration when error
handling for the no addresses situation unintentionally limited the code
to only handling one or zero addresses. The change was made in !1306
(e61b6eb5).

This commit fixes the issue by using the code-structure found in
`load_it_user` and `load_address`, thereby aligning the org-unit address
function with the two others.

The impact of the issue is that we may have been unable to process
org-unit addresses for some users since version 35.1.1, and thus that
the messages for these have been requeued leaving them stuck in the
queue.
Deploying this fix should allow all of these messages to pass the queue
importing the expected addresses.