feat(graphapi): [#64963] expose version collection without authentication
This commit ensures that the `version` collection on GraphQL can be
accessed without authentication, such as with the `curl` call:
```
curl -X POST \
     -H "Content-Type: application/json" \
     -d '{ "query": "query { version { mo_hash mo_version } }" }' \
     http://localhost:5000/graphql/v25
```
This aligns the collection with the service API endpoint `/version/`
which returns the same data (it does in fact use the GraphQL endpoint),
thus resolving an inconsistency in our authentication scheme.

Additionally it allows us to check whether GraphQL works without
depending on whether authentication is configured correctly by using the
above `curl` call as a debugging step.