feat(graphapi): [#64963] expose health collection without authentication
This commit ensures that the `health` collection on GraphQL can be
accessed without authentication, such as with the `curl` call:
```
curl -X POST \
     -H "Content-Type: application/json" \
     -d '{ "query": "query { healths { objects { identifier status } } }" }' \
     http://localhost:5000/graphql/v25
```
This aligns the collection with the service API endpoints `/health/` and
`/health/{identifier}`, which returns the same data
(it does in fact use the GraphQL endpoint),
thus resolving an inconsistency in our authentication scheme.

Additionally it allows us to check whether the health of the application
in yet another way without authentication.