feat(docker): [#63467] add ldapsearch to the container
This commit adds the `ldapsearch` command from `ldap-utils` to the
container, such that we do not have to install it to the host whenever
manual lookups in LDAP are required.

Using `ldapsearch` directly rather than reimplementing the tool within
in the integration itself is deemed desirable as it is a standard tool
with plenty of documentation and examples available online.