test(dataloader): [#65031] modify DataLoader utilisation test for desired behavior
This commit modifies the test introduced in the prior commit to show the
desired behavior, i.e. that our dataloaders are being utilised and are
able to bulk requests as their lookup times are now identical.

The test is now xfailed as the desired behavior described by the test is
not the behavior exhibited by the code.