refactor(client): [#65186] add owner parameter to refresh mutators
This commit adds the `owner` parameter to the existing plain refresh
mutators as it is needed to send refresh events on the new GraphQL-based
event system in MO.

The change is made in preparation of refresh mutators being exposed to
the Jinja template context such that the users of the integration can
requeue events dynamically.

Additionally the arguments of the functions have been swapped such that
the UUIDs (which are required) come first, followed by one of the
optional parameters, either `owner` or `exchange`.