feat(graphql): [#64729] implement rolebinding_delete
This commit adds the `rolebinding_delete` mutator which was missing,
thereby ensuring that rolebindings have a full set of mutators similar
to addresses, classes, employees, engagements, etc.

The mutator seems to have been missing for a long time, as evidenced by
the `# TODO: roles_delete` comment which has been in the code base since
version 12.4.0 (introduced by !1275 (7e69808d)), however at this time
all the role mutators were missing, with create, update and terminate
being introduced in version 23.16.0 (!1914 (b9cfec12)),
23.30.0 (!1918 (7a9725a4)) and 23.31.0 (!1921 (95a1ef39)) respectively.

The replacement of roles with rolebindings in version 36.0.0 with !2346
(25463b28) simply did a 1-to-1 replacement of the old role mutators with
their equivalent rolebinding mutators, and thus did not introducing the
missing mutator.