feat(environment): [#65187] introduce ituser_uuid_to_rolebinding_uuids
This commit introduces the `ituser_uuid_to_rolebinding_uuids` function
which returns a set of rolebinding UUIDs given an ituser UUID.

The function is then used in our newly added role class to group test by
further parametrizing the test such that any changes to an ituser will
resynchronize all rolebindings referencing that ituser.

In theory the groups the ituser is in before should also be
synchronized, but currently we do not handle this for any object types
whatsoever.