feat(environment): [#65187] introduce uuid_to_dn
This commit introduces `uuid_to_dn` to the Jinja template context
offering a way to translate DNs into their corresponding entity UUIDs.

This function is then used in the recently added role to group mapping
test to show that UUIDs stored in the user-key of role classes can be
used to point to the LDAP group to be maintained as opposed to writing
the DN of the group in the user-key of the role class.