feat(environment): [#65187] introduce rolebinding_uuid_to_role_uuids
This commit introduces the `rolebinding_uuid_to_role_uuids` function
which returns a role UUID given a rolebinding UUID.

The function is then used in our newly added role class to group test by
parametrizing the test such that any changes to a role binding will
resynchronize the entire group that rolebinding is now in.

In theory the group the rolebinding is moved from should also be
synchronized, but currently we do not handle this for any object types
whatsoever.