feat(environment): [#65187] introduce itsystem_uuid_to_role_uuids
This commit introduces the `itsystem_uuid_to_role_uuids` function
which returns a set of role UUIDs given an itsystem UUID.

The function is then used in our newly added role class to group test by
further parametrizing the test such that any changes to an itsystem will
resynchronize the entire group that itsystem is referenced by.

In theory the groups the itsystem referenced before should also be
synchronized, but currently we do not handle this for any object types
whatsoever.