test(graphql): [#65974] modify test to disallow engagement termination with leave
This commit modifies the test case introduced in the prior commit to
show the desired behavior of the code rather than the current behavior.

As such the test is marked as `xfail`ed to show that the desired
behavior of the code is not the behavior currently exhibited by the code.

Once the code has been modified to exhibit the desired behavior the
`xfail` can be removed upon which the test should pass.

Note that it may be desirable to develop similar test-cases to this for
other objects that can reference engagements, such as addresses and
itusers.