refactor: [#66087] convert get_ldap_dn to returning None instead of raising
This commit modifies the `get_ldap_dn` method to returning `None` to
signal that the given UUID could not be converted to a DN instead of
raising a `NoObjectsReturnedException`.

We prefer returning `None` to raising as the `None` return is captured
in the typing of the method, and thus forces the caller to handle the
issue via mypy type checking as opposed to exceptions that can easily be
ignored.

What would have been desirable was a way to type that the method raises
a certain exceptions forcing the caller to either handle the exception
or to redeclare it, however PEP-484 intentionally has no syntax for
listing exceptions: https://peps.python.org/pep-0484/#exceptions, and
the suggested solution is to put all possible exceptions in the
doc-string.
Additionally Guido van Rossum is very opposed to the idea of exception
checking using typing, declaring it will not happen:
https://github.com/python/typing/issues/71#issuecomment-87767297

The result being that if one wants type-checked exceptions in Python,
one must use a Rust-style approach to exceptions returning an `Error`
monad: https://doc.rust-lang.org/book/ch09-00-error-handling.html