test: [#66087] split test_extract_unique_ldap_uuids in multiple
This commit splits the `test_extract_unique_ldap_uuids` test in multiple.
One test showing the happy-path behavior, one showing an error when a
single bad user-key is encounter, one showing multiple errors when
multiple bad user-keys are encounter, and two which shows the behavior
when a mix of good and bad user-keys are provided.

Additionally a few tests concerning duplicate entries were added to
cement the behavior of duplicate user-keys.