chore(usernames): [#66058] copy in permutation username generator tests
This commit copies in the `TestUserNameGen` and
`TestUserNameGenPermutation` unit-test classes from the old AD
integration to try to retrain the testing done on the equivalent
classes.

It is regretful that a similar process was not done when the username
generator was introduced to the integration in !13 (d2a6383a), as the
nearly 1000 lines of mixed hypothesis, parametrized and normal tests from
the old AD integration were replaced with 8 non parametrized tests.

This commit is intentionally left broken with unfulfilled dependencies,
imports and such to keep it as close to the original code as possible,
making it easier to trace the changes done in the following commits to
clean it up.