refactor(usernames): [#66058] eliminate max_iterations code
The refactoring made in the prior commit made it so that each iteration
of the while loop would add atleast one character per iteration.
This means that we need to do at most 2 iterations of the while-loop
removing the prior upper-bound set by the `max_iterations` variable at
1000.

This is a big step towards making the code more predictable and allowing
further simplifications.