fix(ldap_amqp): [#66058] only run LDAP event processing when configured
This commit ensures that LDAP events are only being processed if
configured to do so, i.e. only when the `LISTEN_TO_CHANGE_IN_LDAP` flag
is set.

Prior to this commit events would be processed even if not configured to
do so, and thus the `LISTEN_TO_CHANGE_IN_LDAP` would only affect whether
the LDAP event generator was running or not.

This was unexpected as the `LISTEN_TO_CHANGE_IN_MO` flag does not
control whether the MO event generator is running, but rather whether
the integration installs its listeners for MO.