refactor(test): [#65186] introduce write_ldap_api fixture
This commit introduces the `write_ldap_api` fixture which can be used in
test-fixtures that need to be able to write to LDAP even if the
integration is configured as read-only or similar.

At the same time the old `ldap_api` fixture has been reconfigured to
respect the `pytest.mark.envvar`s such that it can be configured as
read-only.