feat(environments): [#65186] add resolve_dar_address to environment
This commit adds the `resolve_dar_address` function to the Jinja
template environment. The function implements behavior equal to the
`split_addresses` functionality in the old AD integration, which was
found within the `ad_writer.py` file.

Thus it now becomes possible to get the `postal_code`, `city` and
`streetname` from a DAR address, making it possible to implement
equivalents to the `unit_postal_code`, `unit_city` and `unit_streetname`
Jinja template context values in the old AD integration.

The function is implemented very differently from the old AD integration
version which extracted the DAR address description string and did
string processing to try to extract the postal code, city and streetname
from it. The new implementation instead asks MO to resolve the DAR address
in DAR itself and passes on the structured data to the integration
eliminating the need for stringification and parsing.