test(api): [#65186] adjust test for no changes to desired behavior
This commit adjusts the test introduced in the prior commit to instead
test for the desired behavior, i.e. that no-changes is produced when no
changes are requested rather than erroreously producing an assertion
error.

The test is marked as xfailed as the current behavior of the program is
not the desired behavior.