test(api): [#65186] adjust test for json formatting to desired behavior
This commit adjusts the test introduced in the prior commit to instead
test for the desired behavior, i.e. that a valid jsonl file is produced
when multiple entities exist in MO rather than an incorrect file.

The test is marked as xfailed as the current behavior of the program is
not the desired behavior.