feat(graphql): [#65186] add supplementary city to DAR addresses
This commit adds a `supplementary_city` as a field to `DARAddress`es
within the `resolve` field of `addresses` in GraphQL.

The field corresponds to the DAR field `"supplerendebynavn"` which
holds the name of a village or other locality, designated by the
municipality to pinpoint an address within a postal code.
It is usually included in the official address of a place.