feat(environments): [#63132] add legacy manager
This commit adds `get_legacy_manager_person_uuids` to the Jinja template
environment, which is very similar to `get_manager_person_uuid` only it
returns the manager that the old AD integration would have returned,
rather than returning the manager that MO provides.

The function is added to allow emulating the behavior of the old AD
integration, which is desirable for adopting the new integration without
changing data in LDAP in the process.
The behavior of the current `get_manager_person_uuid` and the newly
introduced `get_legacy_manager_person_uuids` differs especially when
multiple managers are found on an organisation unit.

The new `get_manager_person_uuid` assumes that a manager can uniquely be
pointed to from the set of potential managers, relying on the data in MO
to be corret.
The newly introduced `get_legacy_manager_person_uuids` on the other hand
gracefully handles multiple conflicting managers by defaulting to the
one with the largest person UUID, unless it is the person themselves, in
which case a manager is sought on the parent organisation unit, unless
it is the root, in which case the manager can manage themselves.

The behavior of the `get_legacy_manager_person_uuids` is designed to
emulate the behavior of the `get_manager_uuid` function on the
`MODataSource`s in the old AD integration, in particular emulating the
`LoraCacheSource`'s `get_manager_uuid` function, which works by doing a
some transformation and inheritence on the `acting_manager_uuid` from the
LoraCache, which itself is implemented by doing a GraphQL lookup followed
by some processing and transformation.

Finding and verifying the old behavior has taken several hours of
studying the old code followed by several hours of tests, but it is hour
belief that the effective behavior of the old `get_manager_uuid` has
been captured in `get_legacy_manager_person_uuids`.

In particular interest for the discussion of equivalence in the
`find_manager` function of `gql_lora_cache_async.py` within the
`sql_export` (LoraCache) integration, which uses `first` to handle
multiple managers.