fix(graphql): [#63132] conditional DAR field lookup
This commit makes field lookups during DAR resolution conditional as
the requested fields may or may not exist depending on whether the
address is an access address, a normal address or a historic address.

The AsyncDARClient underpinning the resolve_dar function and its dar_loader
attempts to load the provided UUID using 4 different types of addresses in
DAR (by default) using a prioritized order defined by the `AddressType` enum
in dar_client.py within FastRAMQPI, the order is:
* "adresser"
* "adgangsadresser"
* "historik/adresser"
* "historik/adgangsadresser"

Thus which fields are available on our `dar_response` depends on which of
these address-types succesfully looked up our DAR UUID.

This is highly problematic as the historic endpoints are marked as
experimental by DAR, and thus may change their fields and behavior whenever.
It is also problematic as the historic endpoints return very little data
compared to their non-historic counterparts.
We should probably never have relied on finding historic addresses, but it is
our interface now, so we cannot easily change the behavior.

Additionally some fields only occur on adresser and not on adgangsadresser,
which is not something that the code was ever designed to handle.

In general it seems the entire AsyncDARClient interface is maldesigned and we
must conditionally check for the existence of all fields we need.

Thus this commit modifies the current `[]` lookup to instead do a `.get`
lookup for all DAR fields to communicate that we do not in fact know if
the requested fields are present or not.