test(graphql): [#63121] assert desired DAR behavior with missing fields
This commit modifies the test introduced in the prior commit to assert
the desired behavior of DAR field resolution on historic addresses where
some fields are missing, i.e. that no error is produced and that the
fields are simply empty.

The test is also marked as `xfail`ed as the desired behavior of the code
is not the current behavior exhibited, something which will be resolved
in the following commit introducing a code change.