feat(discriminator): [#63132] implement legacy discriminator bypass
This commit introduces a new feature flag
`discriminator_legacy_bypass_via_itsystem`, which can be used to bypass
the discriminator system entirely in favor of an attempt at implementing
an emulation of the discriminator behavior found in the old AD integration
by a method entirely different from the one used in the old AD integration.

The old AD integration discriminates multiple AD users by picking the first
user returned by the `Get-ADUser` PowerShell command, which has an undefined
order based on the domain controllers storage engine.

We cannot get this order when connecting to the domain controller via the LDAP
protocol, and thus we cannot simply reproduce the code here. Instead we are
going to use the data left behind by the old integration to attempt to target
the same AD accounts as the old AD integration did.

If no such leftover information can be found, we will instead revert to our
own arbitrary, but deterministic behavior of picking the AD account with the
(alphabetically) lowest DN.
This arbitrary choice should then be cemented when we ourselves create an IT
user which is found as if it was created by the old AD integration.