fix(ldap2mo): [#63132] ensure dry-run termination works
This commit fixes an issue where dry-run termination would fail with a
`JSONDecodeError` instead of raising the `DryRunException` as expected.

The issue was introduced in version 38.4.0 of the integration
(!1387, 57011f6f) when the dry-run flag was being introduced.

The `JSONDecodeError` was triggered from a `TypeError` inside Pydantic
when calling the `jsonable_encoder` from FastAPI. The issue turned out
troublesome to debug, but in the end it was identified to be related to
the `Termination` object, which contains a `mo_class` instance variable
holding a Pydantic class as its value. The `mo_class` instance variable
is required to know which kind of object to terminate, however the
entire construction is a hack to bypass the stack to avoid properly
designing the integration. However the core issue turns out to be that
the Pyndatic class cannot be serialized to JSON and as such would crash.

This commit resolves the issue by excluding `mo_class` from the
serialization, additionally it ensures that `jsonable_encoder` is being
called directly with the Pydantic model instead of with its dictionary
equivalent as described in the FastAPI documentation.

The impact of the issue is minimal as the dry-run can only be triggered
manually, and as such it seems unlikely that anyone has been affected.