chore(graphql): [#65303] add strip_args
This commit adds the `strip_args` resolver transformation function.

Given an asynchronous function and a set of parameter names it returns a
new function which takes the same parameters as the old function with
the provided set of parameter names removed.

This is useful if the parameters to be removed serve no purpose in the
given call context, it can be used for instance to remove pagination
controls from a resolver when a single object is always returned.

The implementation is somewhat similar to `seed_resolver` except
significantly simpler.