feat(graphiql): [#65303] hide deprecated fields in GraphiQL by default
This commit injects a JavaScript snippet into GraphiQL which creates a
custom GraphiQL fetcher to optionally modify the introspection query
used to fetch the GraphQL schema allowing us to fetch the schema without
deprecated fields, presenting a more prestine the schema.

Whether deprecated fields are hidden or not can be controlled by setting
the `?show_deprecated=1` URL parameter, with the parameter set GraphiQL
behaves as normal.