feat(graphql): [#65303] implement GraphQL address response objects directly
This commit replaces the recently added `Address` response object resolver
using a combination of `to_response`, `strip_args` and `seed_resolver`
with a much simpler approach of constructing the `Response` objects
directly via an inline lambda relying on the `Response` object to invoke
the resolver as needed when the object-cache is empty.

This change also ensures that the `uuid` fields can be accessed without
actually having to run the resolver for performance, and allows access
to the `registrations` of a `Response` even if the object itself has
been deleted.

It does however come at the cost of removing all further filtering
options on the response resolvers, however as it stands now there were
in fact none as the `uuids` filter used for all of the resolvers cannot
be used with any other filters, additionally the pagination controls
were explicitly disabled using the `strip_args` decorator and as such no
controls or filters are actually lost.
In the future we may wish to reintroduce the resolver to regain these
abilities, however this should be trivial to change at that time, and
ideally the filtering functionality should probably be introduced on the
`Response` object itself which should expose the capabilities of the
underlying resolver.