fix(graphql): [#65303] visibility_response should use visibilty_uuid not address_type_uuid
This commit fixes a bug where the `visibility_response` on `Address`es
in the GraphQL schema would actually return information about the
`address_type` class instead of information about the `visibilty` class.

The bug was a result of a simple copy-paste error and was unfortunately
not caught by the test-suite or the review process.
The typo introduced in version 47.1.0 of OS2mo (!2631, 9c7de18c) merged
a single day prior to this fix.