fix(graphql): [#65303] lookup org_unit parent responses as org units
This commit fixes an issue where the `parent_response` field on the
`org_unit` collection in the GraphQL schema did not function whatsoever,
as the parent `Response` object was being constructed via an
`OrganisationRead` object instead of via an `OrganisationUnitRead`
object, i.e. the code was trying to read the wrong object type.

The issue was introduced in version 47.4.0 of OS2mo (!2712, f900d342).