feat(graphql): [#67163] enable temporal access on response registrations
This commit enables the temporal access (i.e. `current` and `validities`),
inside of the `registrations` field of `Response` objects, to see the
state of a given object at a specific registration time.

Previously, accessing the temporal state for a specific registration
required the API user to first query the `registrations` field on
`Response` to get the registrations `start` and `end` time, to then
subsequently pass into the top-level collection filter or as an argument
(`registration_time`) to the `current` and `validities` fields.

With this change, it is possible to simply access the temporal fields
within the `Response` `registrations` field itself and thus seamlessly
retrieve the temporal state as it was at the given registration time.

For example, the following query retrieves the state of a facet as
defined by a specific registration:
```graphql
query ReadFacetStateFromResponseRegistrations($uuid: UUID!) {
  facets(filter: {uuids: [$uuid]}) {
    objects {
      registrations {
        current {
          user_key
        }
      }
    }
  }
}
```
Which is equivalent to first querying the `registrations`:
```
query ReadFacetRegistrationStart($uuid: UUID!) {
  facets(filter: {uuids: [$uuid]}) {
    objects {
      registrations {
        start
      }
    }
  }
}
```
And then subsequently querying with a registration time top-level filter
or temporal field argument (`registration_time`):
```graphql
query ReadFacetStateFromFilter($uuid: UUID!, $reg_time: DateTime!) {
  facets(filter: {uuids: [$uuid], registration_time: $reg_time}) {
    objects {
      current {
        user_key
      }
    }
  }
}
```

Furthermore, the `current` and `validities` fields on the `registrations`
support the standard temporal arguments from `Response`, such as `at` on
`current` and `start` / `end` on `validities` ensuring a consistent API.

Additionally various tests has been added to prove the equivalence in
functionality as well as arguments taken.